Disable touch (update LastSeen time) for security sensors
I read the description of the touch function:
The touch () function updates the time the device was last viewed and nothing more. No events or notifications are triggered by touching the device.
After a call, the device displays the new value in the LastUpdate field of the device.

Therefore, for security sensors (with the "sensor" and "motion" alias), the touch function can be omitted. I tried to make changes for myself - and everything works great. The same can be done for a sensor with a type of "switch".

Heartbeat messages will come in other types: LinkQuality, BatteryVoltage, etc.