Update __init__.py
Checked it in my Domoticz. It works!
2020-12-13 23:12:09.296 Status: User: Admin initiated a switch command (225/Lidl Plug/Off)
2020-12-13 23:12:12.917 Status: User: Admin initiated a switch command (225/Lidl Plug/On)
2020-12-13 23:12:15.925 Status: User: Admin initiated a switch command (225/Lidl Plug/Off)
The pushbutton on the  socket does not report a state change back when I manually push it. 
The Link quality of the socket is also popped up as a device which at least gives a value and seems to be displaying the signal quality.