Add temperature sensor to MCCGQ11LM door/windows sensor (#10)
... and not to the older door/window sensor model

Note that the temperature reading is nog very accurate. The temperature_calibration setting can compensate for the higher reading, but not for the slow/limited response to temperature changes. I suspect the response is probably limited by the enclosure of the door/window sensor and it was never really intended to be used as a temperature sensor. Still a nice feature that offers a usable indication of temperature changes.